Modified to work with clap.
The problem was that the original design called for clap to have an
option to get the value from multiple instances of the same option, but
that did not prove to be possible. I would like to see that be possible.

Consider a pull request when you get the chance? See the comment in the
actual code for more detail.