Use a single enum with variants instead of an enum and struct.
I thought it was a serious shortcoming of the original version that
rather than including the required data in my enum variants I stuffed
the enums defining the operation into a struct containing the enum and
the data required for the operation. Problem solved. Maybe now I can
sort out that other little issue...