Use custom parsed_value_of() function in branched clap.
The parsed_value_of() function is of incredibly limited value and I'm
somewhat disappointed in that: it changed my code for getting arguments
from

    let seed = match matches.value_of("seed")
        .and_then(|seed| seed.parse().ok())

to

    let seed = match matches.parsed_value_of("seed")

...each with their attendant match expression error handling nonsense.
I hesitate to submit a pull request for anything so trivial.
Unfortunately, that's pretty much the only useful change I'm able to
make to clap-rs at the moment. I really like this library.